mission-pizza/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# MCP
mcp>=1.0.0

# LLM and Agents
langchain>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.40

# Data Validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# UI
streamlit>=1.28.0
